requirements.txt
pandas
requests
supabase
python-dotenv
openpyxl
xlrd
matplotlib
statsmodels
prophet
tabulate